move ImageDiffType out to separate module
