+++ app/src/lib/stores/app-store.ts
@@ -24,7 +24,12 @@ import { BranchesTab } from '../../models/branches-tab'
 import { CloneRepositoryTab } from '../../models/clone-repository-tab'
 import { CloningRepository } from '../../models/cloning-repository'
 import { Commit } from '../../models/commit'
-import { DiffSelection, DiffSelectionType, DiffType } from '../../models/diff'
+import {
+  DiffSelection,
+  DiffSelectionType,
+  DiffType,
+  ImageDiffType,
+} from '../../models/diff'
 import { FetchType } from '../../models/fetch'
 import { GitHubRepository } from '../../models/github-repository'
 import { Owner } from '../../models/owner'
@@ -70,7 +75,6 @@ import {
   ICompareFormUpdate,
   ICompareToBranch,
   IDisplayHistory,
-  ImageDiffType,
   Popup,
   PopupType,
   PossibleSelections,
