+++ app/src/models/diff/index.ts
@@ -3,3 +3,4 @@ export * from './diff-line'
 export * from './diff-selection'
 export * from './image'
 export * from './raw-diff'
+export * from './image-diff'
