+++ app/src/ui/changes/changes.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react'
 import { Diff } from '../diff'
 import { ChangedFileDetails } from './changed-file-details'
-import { ImageDiffType } from '../../lib/app-state'
-import { DiffSelection, IDiff } from '../../models/diff'
+import { DiffSelection, IDiff, ImageDiffType } from '../../models/diff'
 import { WorkingDirectoryFileChange } from '../../models/status'
 import { Repository } from '../../models/repository'
 import { Dispatcher } from '../../lib/dispatcher'
