+++ app/src/ui/repository.tsx
@@ -10,11 +10,7 @@ import { FilesChangedBadge } from './changes/files-changed-badge'
 import { SelectedCommit, CompareSidebar } from './history'
 import { Resizable } from './resizable'
 import { TabBar } from './tab-bar'
-import {
-  IRepositoryState,
-  RepositorySectionTab,
-  ImageDiffType,
-} from '../lib/app-state'
+import { IRepositoryState, RepositorySectionTab } from '../lib/app-state'
 import { Dispatcher } from '../lib/dispatcher'
 import { IssuesStore, GitHubUserStore } from '../lib/stores'
 import { assertNever } from '../lib/fatal-error'
@@ -22,6 +18,7 @@ import { Account } from '../models/account'
 import { enableNotificationOfBranchUpdates } from '../lib/feature-flag'
 import { FocusContainer } from './lib/focus-container'
 import { OcticonSymbol, Octicon } from './octicons'
+import { ImageDiffType } from '../models/diff'
 
 /** The widest the sidebar can be with the minimum window size. */
 const MaxSidebarWidth = 495
